+++ app/src/main/java/com/example/myapplication/StudentProfileFragment.java
@@ -53,7 +53,7 @@ public class StudentProfileFragment extends Fragment {
     private NavController navController;
     private LinearLayout parentLayout;
     private ScrollView scrollView;
-    private int accountID;
+    private long accountID;
 
     private ProfileModel profileModel;
 
@@ -99,7 +99,7 @@ public class StudentProfileFragment extends Fragment {
         getParentFragmentManager().setFragmentResultListener("accountID", this, new FragmentResultListener() {
             @Override
             public void onFragmentResult(@NonNull String requestKey, @NonNull Bundle result) {
-                accountID = result.getInt("accountID");
+                accountID = result.getLong("accountID");
 
                 // Set the empty text in the student profile screen to the first name of the student
                 txtFirstName.setText(database.getFirstNameFromDatabase(accountID));
@@ -177,7 +177,7 @@ public class StudentProfileFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 Bundle bundle2 = new Bundle();
-                bundle2.putInt("accountID2", accountID);
+                bundle2.putLong("accountID2", accountID);
                 getParentFragmentManager().setFragmentResult("accountID2", bundle2);
                 navController.navigate(R.id.action_studentProfileFragment_to_editStudentProfileFragment);
             }
@@ -188,7 +188,7 @@ public class StudentProfileFragment extends Fragment {
             public void onClick(View v) {
                 // Navigate back to select a user screen
                 Bundle bundle = new Bundle();
-                bundle.putInt("accountID", accountID);
+                bundle.putLong("accountID", accountID);
                 getParentFragmentManager().setFragmentResult("studentgraph", bundle);
                 navController.navigate(R.id.action_studentProfileFragment_to_studentGraph);
             }
@@ -228,7 +228,7 @@ public class StudentProfileFragment extends Fragment {
     /**
      * Checks if user has punch data
      */
-    private boolean hasPunchData(int accountID, MyAppProfileDatabase db) {
+    private boolean hasPunchData(long accountID, MyAppProfileDatabase db) {
         List<PunchModel> punches = new ArrayList<>();
 
         punches = db.getAllPunchesFromProfile(accountID);
@@ -242,7 +242,7 @@ public class StudentProfileFragment extends Fragment {
     /**
      * Populates the graph
      */
-    private void populateGraph(MyAppProfileDatabase database, int accountID) {
+    private void populateGraph(MyAppProfileDatabase database, long accountID) {
         long date;
         List<PunchModel> punches = database.getAllPunchesFromProfile(accountID);
 
@@ -269,7 +269,7 @@ public class StudentProfileFragment extends Fragment {
      * Debug method to populate punch table with 20 data points
      * @Test
      */
-    public void insertFakePunchData(int accountID, MyAppProfileDatabase db) {
+    public void insertFakePunchData(long accountID, MyAppProfileDatabase db) {
         Date date;
         Calendar time;
         time = Calendar.getInstance();
